@@ -26,15 +26,17 @@
         <main.basedir>${project.basedir}</main.basedir>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
-        <jdk.version>1.6</jdk.version>
+        <jdk.version>17</jdk.version>
         <target.dir>target</target.dir>
         <maven.build.timestamp.format>yyyyMMdd</maven.build.timestamp.format>
         <timestamp>${maven.build.timestamp}</timestamp>
         <extraVmArgs/>
 
-        <hazelcast.version>4.0.2</hazelcast.version>
+        <tomcat-version>10.1.16</tomcat-version>
+        <hazelcast.version>5.3.6</hazelcast.version>
+        <servlet.api.version>6.0.0</servlet.api.version>
+        <kryo.version>5.5.0</kryo.version>
 
-        <servlet.api.version>3.0.1</servlet.api.version>
         <junit.version>4.11</junit.version>
         <log4j.version>1.2.12</log4j.version>
 
@@ -57,7 +59,7 @@
         <sonar.language>java</sonar.language>
         <sonar.verbose>true</sonar.verbose>
         <maven.source.plugin.version>2.2.1</maven.source.plugin.version>
-        <kryo.version>5.0.0-RC4</kryo.version>
+
     </properties>
 
     <licenses>
@@ -78,6 +80,11 @@
             <name>Mesut Celik</name>
             <email>mesut@hazelcast.com</email>
         </developer>
+        <developer>
+            <id>blackbluegl</id>
+            <name>Lutz Güttler</name>
+            <email>info@blackbluegl.de</email>
+        </developer>
     </developers>
 
     <organization>
@@ -87,10 +94,7 @@
 
     <modules>
         <module>tomcat-core</module>
-        <module>tomcat7</module>
-        <module>tomcat8</module>
-        <module>tomcat85</module>
-        <module>tomcat9</module>
+        <module>tomcat10</module>
     </modules>
 
     <repositories>
@@ -162,12 +166,12 @@
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
             <artifactId>httpclient</artifactId>
-            <version>4.3.6</version>
+            <version>4.5.13</version>
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>javax.servlet</groupId>
-            <artifactId>javax.servlet-api</artifactId>
+            <groupId>jakarta.servlet</groupId>
+            <artifactId>jakarta.servlet-api</artifactId>
             <version>${servlet.api.version}</version>
             <scope>provided</scope>
         </dependency>
